web-service/backend_fastapi/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymysql==1.1.0
sqlalchemy==2.0.23
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
pydantic==2.4.2
websockets==11.0.3
requests==2.31.0
python-dotenv==1.0.0